+++ src/Panel/index.js
@@ -1,0 +1,6 @@
+import Panel from './Panel'
+import PanelHeader from './PanelHeader'
+import PanelBody from './PanelBody'
+import './index.less'
+
+export { Panel, PanelHeader, PanelBody }